// src/data/api/types/store-daily-progress-check-res.ts
// Generated by https://quicktype.io

export interface StoreDailyProgressCheckRes {
    message: string;
    data:    Data;
}

export interface Data {
    id:                  string;
    judgement:           string;
    judgement_3d:        string;
    judgement_2d:        string;
    shift:               string;
    pic:                 string;
    inspection_date:     Date;
    part_weight_qis:     string;
    lot_production:      string;
    label_no:            number;
    accept_sample_time:  string;
    measure_sample_time: string;
    actual_part_weight:  number;
    part_id:             string;
    machine_id:          string;
    updated_at:          Date;
    created_at:          Date;
    checked_by:          null;
    deletedAt:           null;
}